fixed transactions redux
removed unused transactions folder, refactored selectors names, organized files
