+++ src/components/dashboard/header/Header.js
@@ -1,6 +1,5 @@
-import { getModalState } from 'redux/transactions/selectors';
 import { toggleLogOutModal } from 'redux/modal/ModalSlice';
-import { getModalTypeState } from 'redux/modal/selectors';
+import { selectModalState, selectModalTypeState } from 'redux/modal/selectors';
 import Modal from '../../Modal/Modal';
 import Logout from '../../Logout/Logout';
 import { HeaderStyled } from './Header.styled';
@@ -8,10 +7,10 @@ import { HeaderStyled } from './Header.styled';
 export const Header = () => {
   const { useDispatch, useSelector } = require('react-redux');
   const dispatch = useDispatch();
-  const modalType = useSelector(getModalTypeState);
-  const isModalOpen = useSelector(getModalState);
+  const modalType = useSelector(selectModalTypeState);
+  const isModalOpen = useSelector(selectModalState);
 
-    return (
+  return (
     <HeaderStyled>
       <h2>Логотип, имя, Exit</h2>
       <button type="button" onClick={() => dispatch(toggleLogOutModal())}>
@@ -19,8 +18,8 @@ export const Header = () => {
       </button>
 
       {modalType === 'modal/toggleLogOutModal' && isModalOpen && (
-          <Modal children={Logout()} />
+        <Modal children={Logout()} />
       )}
     </HeaderStyled>
   );
-}
+};
