+++ src/redux/modal/selectors.js
@@ -1,2 +1,3 @@
-export const getModalTypeState = state => state.modal.modalType;
-export const getModalState = state => state.modal.isModalOpen;
+export const selectModalTypeState = state => state.modal.modalType;
+export const selectModalState = state => state.modal.isModalOpen;
+// export const selectModalState = state => state.modal.isModalOpen;
