+++ src/redux/transactionsRedux/transactionsSelectors.js
@@ -1,2 +1,5 @@
 export const selectorTransactions = state => state.transactions.transactions;
 
+export const selectError = state => state.transactions.error;
+
+export const selectIsLoading = state => state.transactions.isLoading;
